refactor(lifecycle): extract risk policy
